Fix cuDNN batchnorm gradients in testmode
The cuDNN batchnorm implementation was calculating an incorrect gradient
in testmode (reported in FluxML/Flux.jl#2179). This was due to relying
on cudnnBatchNormalizationBackward, the train mode implementation. Batch
normalization in test mode is a different algorithm, and since cuDNN
doesn't provide a test mode implementation, the fix was to calculate the
gradients ourselves.